directly convert the metadata to something useful in `read_metadata`
